Add Deluge page and update environment variable handling for API
