+++ .env.example
@@ -4,4 +4,4 @@ DELUGE_URL=
 DELUGE_PASSWORD=
 # (Optional) Request timeout in ms
 DELUGE_TIMEOUT=30000
-NEXT_PUBLIC_BASE_URL=
+DELUGE_NEXT_BASE_URL=
