+++ docker-compose.yml
@@ -9,4 +9,4 @@ services:
       - DELUGE_URL=${DELUGE_URL}
       - DELUGE_PASSWORD=${DELUGE_PASSWORD}
       - DELUGE_TIMEOUT=${DELUGE_TIMEOUT}
-      - NEXT_PUBLIC_BASE_URL=${NEXT_PUBLIC_BASE_URL}
+      - DELUGE_NEXT_BASE_URL=${DELUGE_NEXT_BASE_URL}
